+++ thruster-async-await/Cargo.toml
@@ -22,9 +22,7 @@ thruster_error_handling = [
 
 [dependencies]
 futures-legacy = { version = "0.1.23", package = "futures" }
-futures-preview = "0.3.0-alpha.16"
+futures-preview = { version = "0.3.0-alpha.16", features = ["compat"] }
 futures-util = "0.2.1"
 thruster-core = { version = "0.7", path = "../thruster-core" }
-tokio = { git = "https://github.com/tokio-rs/tokio", branch = "master" }
-tokio-futures = { git = "https://github.com/tokio-rs/tokio", branch = "master", features = ["async-await-preview"] }
-
+tokio = { version = "0.1.20" }
