+++ thruster-async-await/src/lib.rs
@@ -6,10 +6,9 @@ use futures_legacy::{Future as FutureLegacy};
 use thruster_core::context::Context;
 use thruster_core::request::{RequestWithParams};
 use thruster_core::route_parser::{MatchedRoute};
-use thruster_core::errors::Error;
 
 pub fn resolve<R: RequestWithParams, T: 'static + Context + Send>(context_generator: fn(R) -> T, mut request: R, matched_route: MatchedRoute<T>) -> impl FutureLegacy<Item=T::Response, Error=io::Error> + Send {
-  use tokio_futures::compat::into_01;
+  use futures::future::{FutureExt, TryFutureExt};
 
   request.set_params(matched_route.params);
 
@@ -30,5 +29,7 @@ pub fn resolve<R: RequestWithParams, T: 'static + Context + Send>(context_genera
     Ok(ctx.get_response())
   };
 
-  into_01(context_future)
+  context_future
+    .boxed()
+    .compat()
 }
