+++ thruster-core/Cargo.toml
@@ -31,6 +31,6 @@ serde_derive = "1.0.24"
 smallvec = "0.6.2"
 templatify = "0.2.3"
 time = "0.1"
-tokio-codec = { git = "https://github.com/tokio-rs/tokio", branch = "master" }
-tokio-io = { git = "https://github.com/tokio-rs/tokio", branch = "master" }
+tokio-codec = { version = "0.1.1" }
+tokio-io = { version = "0.1.12" }
 thruster-core-async-await = { version = "0.7", path = "../thruster-core-async-await", optional = true }
